@@ -8,38 +8,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.7",
-    "@angular/cdk": "^6.3.2",
-    "@angular/common": "^6.0.7",
-    "@angular/compiler": "^6.0.7",
-    "@angular/core": "^6.0.7",
-    "@angular/forms": "^6.0.7",
-    "@angular/http": "^6.0.7",
-    "@angular/material": "^6.3.2",
-    "@angular/platform-browser": "^6.0.7",
-    "@angular/platform-browser-dynamic": "^6.0.7",
-    "@angular/pwa": "^0.6.8",
-    "@angular/router": "^6.0.7",
-    "@angular/service-worker": "^6.0.7",
-    "@ngxs/router-plugin": "^3.1.4",
-    "@ngxs/storage-plugin": "^3.1.4",
-    "@ngxs/store": "^3.1.4",
-    "angularfire2": "^5.0.0-rc.11",
-    "core-js": "^2.5.7",
-    "firebase": "^5.2.0",
+    "@angular/animations": "^7.2.15",
+    "@angular/common": "^7.2.15",
+    "@angular/compiler": "^7.2.15",
+    "@angular/core": "^7.2.15",
+    "@angular/fire": "^5.1.3",
+    "@angular/forms": "^7.2.15",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^7.3.7",
+    "@angular/platform-browser": "^7.2.15",
+    "@angular/platform-browser-dynamic": "^7.2.15",
+    "@angular/pwa": "^0.13.9",
+    "@angular/router": "^7.2.15",
+    "@angular/service-worker": "^7.2.15",
+    "@ngxs/storage-plugin": "^3.4.3",
+    "@ngxs/store": "^3.4.3",
+    "core-js": "^2.6.8",
+    "firebase": "^6.1.0",
     "hammerjs": "^2.0.8",
-    "moment": "^2.22.2",
-    "rxjs": "^6.2.1",
-    "zone.js": "^0.8.26"
+    "rxjs": "^6.5.2",
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.6.8",
-    "@angular/cli": "^6.0.8",
-    "@angular/compiler-cli": "^6.0.7",
-    "@angular/language-service": "^6.0.7",
-    "@types/node": "^10.5.2",
-    "codelyzer": "^4.4.2",
-    "tslint": "^5.10.0",
-    "typescript": "^2.7.2"
+    "@angular-devkit/build-angular": "^0.13.9",
+    "@angular/cdk": "^7.3.7",
+    "@angular/cli": "^7.3.9",
+    "@angular/compiler-cli": "^7.2.15",
+    "@angular/language-service": "^7.2.15",
+    "@types/node": "^12.0.3",
+    "codelyzer": "^5.0.1",
+    "tslint": "^5.16.0",
+    "typescript": "^3.2.4"
   }
 }
