@@ -1,42 +1,79 @@
-import { IObject } from "@lib/object.library";
-import { TStore } from "@dbase/data/data.interface";
-
-export const COLLECTION = {			// collection starts with upper-case, document with lower-case
-	Log: 'log',
-	Register: 'register',
-	Member: 'member',
-	Attend: 'attend',
-	Client: 'client',
+export enum COLLECTION {
+	log = 'log',
+	member = 'member',
+	attend = 'attend',
+	client = 'client',
+	admin = 'admin',
+	device = 'device',
 }
 
-export const STORE: IObject<TStore> = {
-	profile: 'profile',
-	attend: 'attend',
-	provider: 'provider',
-	class: 'class',
-	event: 'event',
-	price: 'price',
-	plan: 'plan',
-	schedule: 'schedule',
+export enum STORE {
+	profile = 'profile',
+	payment = 'payment',
+	attend = 'attend',
+	provider = 'provider',
+	alert = 'alert',
+	class = 'class',
+	event = 'event',
+	price = 'price',
+	plan = 'plan',
+	schedule = 'schedule',
+	calendar = 'calendar',
+	location = 'location',
+	instructor = 'instructor',
+	message = 'message',
+	bonus = 'bonus',
+	gift = 'gift',
+	span = 'span',
+	default = '_default_',
+	config = '_config_',
+	schema = '_schema_',
+	register = 'register',
+	migrate = 'migrate',
+	status = 'status',
+	local = '@config@',
+	log = 'log',
 }
 
-export const enum FIELD {					//  common Field names
-	effect = '_effect',
-	expire = '_expire',
-	modify = '_modify',
-	hidden = '_hidden',
-	disable = '_disable',
+export enum MEMBER {
+	profile = 'profile',
+	payment = 'payment',
+	diary = 'diary',
+	gift = 'gift',
+	status = 'status',
+}
+
+export enum FIELD {								// common Field names
+	effect = '_effect',							// valid when greater-than-or-equal-to
+	expire = '_expire',							// valid when less-than
+	update = '_update',							// date-time modified
+	create = '_create',							// date-time created
+	access = '_access',							// date-time accessed
+	hidden = '_hidden',							// valid, but not displayed to User
+	disable = '_disable',						// displayed, but greyed to User
+	stamp = 'stamp',
+	note = 'note',
+	date = 'date',
 	id = '_id',
 	uid = 'uid',
-	key = '_key',
 	store = 'store',
 	type = 'type',
-	memberId = 'memberId'
+	icon = 'icon',
+	key = 'key',
 };
 
-/** a list of Fields on which to Filter Documents, when determining Type-2 delta-inserts */
-export const FILTER: IObject<string[]> = {
-	[COLLECTION.Client]: [FIELD.store],
-	[COLLECTION.Member]: [FIELD.store, FIELD.type, FIELD.uid],
-	[COLLECTION.Attend]: [FIELD.store, FIELD.type, FIELD.uid],
+export enum CONNECT {
+	connect = 'connect',						// connected, but not logged-in
+	active = 'active',							// logged-in, browser has focus
+	behalf = 'behalf',							// logged-in on behalf of Member
+	inactive = 'inactive',					// logged-in, browser lost focus
+	offline = 'offline',						// logged-in, network off air
+	logout = 'logout',							// logged-out, browser lost focus
+}
+
+export enum BONUS {
+	gift = 'gift',
+	week = 'week',
+	month = 'month',
+	sunday = 'sunday',
 }
