@@ -1,61 +1,63 @@
-import { State, Action, StateContext, Selector, NgxsOnInit } from '@ngxs/store';
-import { SLICE } from '@dbase/state/store.define';
-import { IStoreState, IStoreDoc } from '@dbase/state/store.define';
-import { SetMember, DelMember, TruncMember } from '@dbase/state/store.define';
-import { currStore } from '@dbase/state/store.state';
+import { State, Action, StateContext, NgxsOnInit } from '@ngxs/store';
+import { TStateSlice, SLICE } from '@dbase/state/state.define';
+import { IStoreMeta } from '@dbase/data/data.schema';
+import { SetMember, DelMember, TruncMember, filterState } from '@dbase/state/state.action';
 
-import { FIELD, STORE } from '@dbase/data/data.define';
-import { sortKeys } from '@lib/object.library';
+import { FIELD } from '@dbase/data/data.define';
+import { asArray } from '@lib/array.library';
 import { dbg } from '@lib/logger.library';
 
-@State<IStoreState>({
+@State<TStateSlice<IStoreMeta>>({
 	name: SLICE.member,
 	defaults: {}
 })
 export class MemberState implements NgxsOnInit {
-	private dbg: Function = dbg.bind(this);
+	private dbg = dbg(this);
 
-	constructor() { }
+	constructor() { this.init(); }
 
-	ngxsOnInit(_ctx: StateContext<IStoreState>) { this.dbg('init:'); }
+	ngxsOnInit(_ctx: StateContext<TStateSlice<IStoreMeta>>) { this.init(); }
+
+	private init() {
+		this.dbg('init:');
+	}
 
 	@Action(SetMember)
-	setStore({ patchState, getState }: StateContext<IStoreState>, { payload }: SetMember) {
+	setStore({ getState, setState }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: SetMember) {
 		const state = getState() || {};
-		const store = this.filterMember(state, payload);
 
-		store.push(payload);										// push the changed MemberDoc into the Store
-		state[payload.store] = store;
-		this.dbg('setMember: %j', payload);
-		patchState({ ...state });
+		asArray(payload).forEach(doc => {
+			const store = doc[FIELD.store];
+
+			state[store] = filterState(state, doc);
+			state[store].push(doc);							// push the changed MemberDoc into the Store
+
+			if (debug) this.dbg('setMember: %j', doc);
+		})
+
+		setState({ ...state });
 	}
 
 	@Action(DelMember)
-	delStore({ patchState, getState }: StateContext<IStoreState>, { payload }: DelMember) {
+	delStore({ getState, setState }: StateContext<TStateSlice<IStoreMeta>>, { payload, debug }: DelMember) {
 		const state = getState() || {};
-		const store = this.filterMember(getState(), payload);
+		const segment = FIELD.store;
 
-		state[payload.store] = store;
-		this.dbg('delMember: %j', payload);
-		patchState({ ...state });
+		asArray(payload).forEach(doc => {
+			const slice = doc[segment];
+			state[slice] = filterState(state, doc, segment);
+
+			if (state[slice].length === 0)
+				delete state[slice];
+			if (debug) this.dbg('delMember: %j', doc);
+		})
+
+		setState({ ...state });
 	}
 
 	@Action(TruncMember)
-	truncStore({ setState }: StateContext<IStoreState>) {
-		this.dbg('truncMember');
+	truncStore({ setState }: StateContext<TStateSlice<IStoreMeta>>, { debug }: TruncMember) {
+		if (debug) this.dbg('truncMember');
 		setState({});
-	}
-
-	/** remove an item from the Member Store */
-	private filterMember(state: IStoreState, payload: IStoreDoc) {
-		const curr = state && state[payload.store] || [];
-
-		return [...curr.filter(itm => itm[FIELD.id] !== payload[FIELD.id])];
-	}
-
-	/** Selectors */
-	@Selector()											/** Get current (un-expired) documents */
-	static getMember(state: IStoreState) {
-		return currStore.bind(this, state);
 	}
 }
