@@ -1,6 +1,8 @@
 import { NgModule } from '@angular/core';
+import { DragDropModule } from '@angular/cdk/drag-drop';
 
 import { MATERIAL_SANITY_CHECKS } from '@angular/material';
+import { MatDialogModule } from '@angular/material';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MatToolbarModule } from '@angular/material/toolbar';
@@ -11,25 +13,40 @@ import { MatDatepickerModule } from '@angular/material/datepicker';
 import { MatNativeDateModule } from '@angular/material';
 import { MatExpansionModule } from '@angular/material/expansion';
 import { MatDividerModule } from '@angular/material/divider';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { MatTabsModule } from '@angular/material/tabs';
+import { MatBadgeModule } from '@angular/material/badge';
+import { MatSnackBarModule, MAT_SNACK_BAR_DEFAULT_OPTIONS } from '@angular/material/snack-bar';
 
 const modules = [
-  MatToolbarModule,
-  MatMenuModule,
-  MatButtonModule,
-  MatIconModule,
-  MatCheckboxModule,
-  MatCardModule,
-  MatDatepickerModule,
-  MatNativeDateModule,
-  MatExpansionModule,
-  MatDividerModule,
-  MatSnackBarModule,
+	MatToolbarModule,
+	MatDialogModule,
+	MatMenuModule,
+	MatButtonModule,
+	MatIconModule,
+	MatCheckboxModule,
+	MatCardModule,
+	MatDatepickerModule,
+	MatNativeDateModule,
+	MatExpansionModule,
+	MatDividerModule,
+	MatSnackBarModule,
+	MatTabsModule,
+	MatBadgeModule,
 ]
 
 @NgModule({
-  imports: [...modules],
-  exports: [...modules],
-  providers: [{ provide: MATERIAL_SANITY_CHECKS, useValue: false }],
+	imports: [
+		DragDropModule,
+		...modules,
+	],
+	exports: [
+		DragDropModule,
+		...modules
+	],
+	providers: [
+		{ provide: MATERIAL_SANITY_CHECKS, useValue: false },
+		{ provide: MAT_SNACK_BAR_DEFAULT_OPTIONS, useValue: { duration: 5000 } },
+	],
+	declarations: [],
 })
 export class MaterialModule { }
