@@ -1,3 +1,20 @@
 export const environment = {
-  production: true
+  production: true,
+
+  firebase: {
+    app: {
+      apiKey: "AIzaSyCQt465v9SVXfKE2gi5kHxWgExYyiXbQXQ",
+      authDomain: "whitefire-dev.firebaseapp.com",
+      databaseURL: "https://whitefire-dev.firebaseio.com",
+      projectId: "whitefire-dev",
+      storageBucket: "whitefire-dev.appspot.com",
+      messagingSenderId: "919087826501",
+    },
+
+    config: {
+      name: 'whiteFire',
+      automaticDataCollectionEnabled: true,
+      // timestampsInSnapshots: true,
+    }
+  }
 };
